mb_app/requirements.txt
flask
flask_restful
mysql-connector
nose2
coverage